WL#11840: add support for standard X DevAPI resultsets
The current approach used by Connector/Node.js with regards to handling
result-set data and metadata is not the most common since it mixes two
different kinds of asynchronous constructs, callbacks to handle result-set
data and Promises to ensure the operation is finished and the application
can check its status. The reason for this is that it helps to avoid
buffering data at the client layer, which can be a problem when working
with large datasets.

So, this patch introduces an additional set of constructs based on the
official X DevAPI standard which buffer result-set data and metadata and
allows applications to use only a given asynchronous construct (Promises)
and consume them using pull-based cursors directly in the Result instance
via methods such as "fetchOne()" and "fetchAll()".