WL#12627: add support for setting the server default schema
The server will now automatically switch to any default schema provided
as part of the connection options. If the schema does not exist, an error
will be thrown.

This is a breaking change wherein previously the schema was being created
automatically, even though there was no switch on the server side.