BUG#28977649: TLS errors not being returned
ce53dcd introduced a regression that prevented TLS errors to bubble up
to the application level. The behavior was ensured by a docker-based
automated test which was not really well maintained and was hard to run.

The docker-based test suite has been updated to mitigate similar issues
in the future.