BUG#29179372: wait for pool conections to become available
Connection attemps were not being queued when a specific and finite
value was provided for queueTimeout when creating a connection pool.
Instead, the client would always throw a timeout error. This has been
fixed and the client will now retry acquiring a connection from the
pool within a sliding window until it reaches the value of queueTimeout.