WL#12484: support different session reset behaviors
The new version of the X Protocol introduces changes to the behavior of
Mysqlx.Session.Reset messages. Right now, it is possible to keep connections
open without the need for re-authentication. This allows, among other things,
to acquire idle connections from a pool faster. It should happen on-demand and
is not supported on older server versions, which means the client must be
flexible enough to check for server support fall back to the previous behavior
if that is the case.