WL#12483: add support for connection attributes
When creating an X DevAPI legacy session or pooling client, an application
can set a list of custom named parameters and respective values that get
attached to the underlying connection.

Additionally, the client also specifies a list of standard attributes that
are attached to the connection by default.

These attributes are stored in a MySQL server using the following
`PERFORMANCE_SCHEMA` tables:

  - `session_account_connect_attrs`
  - `session_connect_attrs`