BUG#29392088: prioritize idle connections in a pool
When a connection pool is not full and contains unused connection slots,
those connection slots are being prioritized instead of re-using idle
connections. This has been optimized in a way that idle connections,
which faster to spawn since they only need to be reset (don't require
re-authentication) are now prioritized over unused ones.