BUG#29427271: disable timeout after successful connection
The socket timeout was not being disabled after a successful connection.
That meant that after a period of socket inactivity as defined by the
`connect-timeout` property, subsequent network calls were being prevented
and an error was being thrown.

This should not be the case, since the timeout only applies to the
connection establishment process and not further client-erver interaction.