WL#12746: add support for TLS version selection
Applications can now provide a list of TLS protocol versions to pass to
the underlying Node.js OpenSSL library. Starting with Node.js v10.6.0, a
range is defined between the oldest and the newest TLS version that are
available in that list. In older Node.js versions, the newest TLS version
is selected as the baseline. By default, the client relies on
tls.DEFAULT_MAX_VERSION (TLSv1.3 with Node.js >=v12.0.0 and TLSv1.2 on
other Node.js versions).