WL#13358: Collection.getOne() optimizations
Collection.getOne() was implemented in a way that required the underlying
lookup expression to be parsed for each and every call. This is totally
unnecessary since we are able to make the expression deterministic by
using a placeholder and assign a value to that placeholder through late
binding, avoiding any additional parsing.

A statement was also created for each and every call. Once again, this
becomes unnecessary when using a deterministic expression. So, by creating
the statement only once, we can also take advantage of server-side prepared
statements, which leads to additional performance improvements.