BUG#30030159: fixed-length column padding based on charset
The padding character to use when filling values with a length lower
than a fixed-length datatype column was being selected based on the
type of the value itself. This is wrong, and the character should be
selected based on the charset that maps to the column collation.

For columns with a "binary" charset, the padding character should be
"0x00" whereas for other charsets, the character should be a space
character as defined by that charset.