BUG#30401432: unexpected error with Collection.existsInDatabase()
Collection.existsInDatabase() was sending an X Plugin 'list_object'
command containing an unknown 'filter' property in the argument list.
Since 8.0.19, the server is validating all the options provided as
arguments, which caused the operation to start failing.

The 'filter' property has been replaced by the 'pattern' property, which
is, in fact, the correct property for that specific purpose.

Additionally, both Collection.existsInDatabase() and
Table.existsInDatabase() were unable to make a distinction between
tables and collections which means both were returning 'true' in the
presence of one database object with the given name, regardless of its
nature (regular table or collection). This has been fixed, and now
Collection.existsInDatabase() returns 'true' only and if only there is a
collection with the given name in the database and
Table.existsInDatabase() returns 'true' only and if only there is a
regular table with the given name in the database. Otherwise, both
methods return 'false'.