BUG#30423556: fix unexpected error with where()
Building a relational table statement with the CRUD API and providing an
undefined filtering criteria expression resulted in an expected error
when calling "execute()".

This has been fixed and behavior for undefined filtering criteria
expressions using "where()" is now the same as the behavior for empty
filtering criteria expressions using "where('')".

Additionally, the error messages for the document store CRUD API
counterpart have been updated for the sake of consistency.