BUG#30401962: re-introduce getAffectedItemsCount() for session.sql()
The getAffectedItemsCount() has been removed from the API of result-sets
returned by session.sql(). This was a mistake that introduced a
regression. The method has been added to the API once again.

Additionally, the getAffectedRowsCount() method has been deprecated
since it is not part of the official X DevAPI standard.