tests: fix breaking expectations for BUG#30162858
The patch for addressing the issues described by BUG#30162858 changes
the client-side type for a MySQL VARBINARY from a JavaScript String into
a Node.js Buffer. A couple of compatibilty tests that have expectations
about this specific client-side type were not updated accordingly, which
means they were now failing.

This patch updates the tests to match the current expected behavior.