tests: update collation tests due to X Plugin breaking change
In MySQL 8.0.20, the X Plugin introduces a breaking change that fixes an
issue causing an incorrect value to be assigned to the
collation_connection session variable.

This issue is reproducible when using a custom collation (non utfmb4)
for the collation_server variable as well. Some tests were assuming the
incorrect value returned by the X Plugin, part of the column metadata,
as being correct which means the assertions started to fail when the
change was introduced.