tests: enable functional test for TIME datatype
The functional test for the MySQL TIME datatype was disabled because we
believe there was some issue with the MySQL server or X Plugin and had
incorrect expectations about the result-set. This wasn't the case, and
the both the test and the setup were perfectly fine.

Additionally, we also assumed that the X Plugin didn't properly encode
fractional seconds in the result-set. This wasn't also the case. The
issue was that the datatype was being incorrectly defined without the
proper precision.