misc: remove duplicate protobuf stubs
There were a few duplicated protobuf stub files in the source tree. This
was becase if it came the time to extract the Expression Parser
component into a standalone Node.js module/package, it would already
contain the necessary stubs to operate.

However, there is currently no plan to do that and the duplicate files
are simply increasing the size of the final package, so for now we
decided to remove them.