BUG#31037211: empty result sets can be consumed
Result.nextResult() was returning "false" when the result set was empty.
This was causing issues with common iterator patterns such as:

do {
  // do something
} while (res.nextResult())

In this case, if the underlying query was yielding multiple result sets,
the later ones could not be consumed at all if some result sets consumed
before were empty (https://stackoverflow.com/q/60544201/3235909).