BUG#30922711: provide useful details on Column.getType()
Column.getType() was returning an internal code id for the type of
column being processed. This was not useful in a way that it wasn't
easy for the application to map that code into valuable information
unless by using the protobuf stub buried deep into the project source
tree. Also, this was in a way, a breaking change that slipped through
the cracks on an older release.

This is now fixed, and the client first maps that code into a proper
type string that gets sent back as part of the column metadata object.
This type string corresponds to an X DevAPI specific type, and not
really a SQL data type or a X Protocol field type.