misc: deprecate object callback API for SqlExecute
SqlExecute provides an API that supports an object containing properties
for the data and metadata callback handlers. This is not in line with
similar APIs such as TableSelect and becomes less relevant with the
introduction of the cursor-based result set API.

This API is being deprecated in favour of the other existing push-based
callback API using different functional parameters for each handler
("execute(dataHandler, metadataHandler)").