misc: fix failing test on later Node v10 releases
One of the TLS selection tests was failing on later Node.js v10
releases. This was an issue with a test pre-condition and not an issue
with the implementation. The implementation relies on the API available
in the core platform, which means that on older Node.js versions it will
only use the minimum common denominator between the selected TLS
versions and on newer Node.js versions it will rely on a range-based
selection.

The specific test should only run on Node.js versions below 10.16.0 and
a new test was introduced that runs only on versions between 10.16.0
(inclusive) and 12.0.0 (exclusive).