WL#13854: re-structure the functional test suite
The functional test suite was structured in a way that made extremely
restrictive and hard to run across different platforms. UNIX sockets
were being privileged over TCP connections, which created some problems
and lead to a lot of hand-waving while running the tests using servers
on different hosts or using Windows.

A small subset of these tests were coupled to a weird setup of Docker
containers, and they were pretty much impossible to run out-of-the-box.

This patch changes the test suite to run using TCP connections by
default and makes UNIX socket connections a specific test concern that
gets verified only under the appropriate platforms (such as Linux and
macOS).

The test suite is now structured in a way that allows to autonomously
run all the possible and valid tests within a given environment, even
those that require server configuration.

The tests where that is simply not possible are able to run using a new
and improved setup of containers which only requires the user to have
Docker installed in their machine.

Some tests and test suites were moved, new ones were extracted or
introduced and all flakyness was removed, leading to an increase of the
code coverage.

Additionally, a couple of tiny changes were made in the implementation,
in order to account for some corner-cases that weren't being verified
in the old test suite. Worth mentioning the fact that failover was not
kicking in for connections where the socket was abrutely interrupted
without a proper error context.