BUG#31418813: restrict offset() to TableSelect and CollectionFind
The "offset()" method is available on APIs such as CollectionModify,
CollectionRemove, TableUpdate and TableDelete, however the X Plugin
currently does not support the specific boundary for those kind of
operations. This leads to an error state where the server forcefully
closes the client connection.

To avoid this situation and to prevent any other source of confusion,
this kind of functionality should not be exposed by the X DevAPI. So,
the "offset()" method has been removed from all the interfaces mentioned
above.