BUG#30845366: decode column metadata on-demand
Binary column metadata was being decoded into utf8 unnecessarily at the
client level, even if the metadata was not used by the application.
Additionally, the decoding happened in a way where new memory was being
allocated for each property.

This patch introduces the changes required to use a lazy decoding
strategy for column metadata fields where it only happens when the
application effectively uses the Column API. Also the decoding now
uses a view on top of the already allocated memory for the binary
content.

As part of this patch, the main work queue handler for server messages
has been refactored to remove dead code, increase readability and
simplify the auxiliary data structures.