tests: allow to run unit test suite in isolation
The way the previous test suite was structured resulted in a byproduct
where, luckily, the client-side authentication plugins were being
registered beforehand. With the recent changes introduced in the test
suite, this ceased to be the case.

As a consequence, the unit test suite couldn't run in isolation because
some tests of the Session component required the authentication plugins
to be available.

This patch introduces the changes required to load the authentication
plugins before the affected tests are run individually or in the scope
of the unit test suite.