tests: enable test with inconsistent X Plugin behavior
There was an X Plugin issue creating an inconsistency between the error
codes returned by the server when a unique key constraint was violated
while performing an upsert. To prevent any source of confusion and the
resulting failure, the test suite was skipping a valid test for that
behavior.

The X Plugin has since then fixed the behavior and there's no need to
keep the test disabled on our end.