WL#13587: improve logging framework and debug mode
Although the debug mode worked fine, it had a few issues and limitations
which made it not much useful, and at some point, even making it harmful.

One of the issues was the fact that, a lot of debug-related work was
being performed even when an app was not running in debug mode. As an
example, All the content of X Protocol protobuf messages was being
decoded unnecessarily, for debug purposes, on every exchange with the
server (as described by BUG#31584269).

Additionally, it was solely relying on the API available by the 3rd-party
protobuf library, which itself has a lot of limitations when it comes to
convert data from binary format into language-specific types.

Also, the logging aspect of it was also pretty limited in the way the
sections were organized and structured and with the impossibility to
have proper utf8 textual content for binary-encoded protocol data. This
meant it was hard to apply any kind of filter and to access the
effective property values composing each protocol message. In one
specific case, there were duplicate messages being logged
(BUG#31729043).

All these issues have been addressed but that required a big redesign
of the Protocol adapter architecture. Each X Protocol protobuf message
is now encapsulated under a proper wrapper, which in turn, contains the
logic to handle the data as desired at the application level. A specific
set of inbound and outbound handlers have been defined to ensure a more
natural structure of the log sections and allow for more interesting
filtering patterns.

The logger itself required some small changes to ensure nothing relevant
was happening whilst an application was not running in debug mode.

This redesign also implicitly introduced the changes required to address
additional issues, not related to logging or debug mode in any capacity,
such as the the ones described by:

- BUG#31734504
- BUG#31709879