tests: fix test forcing default port
The automated functional test for userless connections with a connection
string is always using the default X Plugin port (33060). This makes the
test fail on enviroments where the server is not available on such a
port.

The test is now using the usual custom configuration built based on the
set of default and environment variable values available by the time it
runs.