BUG#31686805: non-BIGINT value in BIGINT column is not decoded
If a table contains a BIGINT column with any other kind of INT values,
the values in the result set are not properly decoded when they are
retrieved by some SQL statement or CRUD query.

This happens because the "readZigzagVarint64String()" method of the
binary decoder available in the google-protobuf library is not capable
of decoding values that don't lose precision.

This can be fixed by first attempting to decode a value as a JavaScript
number, and only if there is a chance of loosing precision, reset the
decoder cursor and attempt to decode the value a a JavaScript string.