WL#14201: handle connection close notifications
MySQL 8.0.23 server started sending asynchronous global notifications
(via the X Plugin) before closing an existing client connection due to
one of the following reasons:

- the connection is idle (according to "mysqlx_wait_timeout" and/or
"mysqlx_read_timeout")
- the connection has been killed in a different session
- the server itself is shutting down

The X DevAPI client is now writing all these notifications in the
NODE_DEBUG log and ensures that when a connection has been closed and
tries to be re-used by an existing session, an error, pointing to the
specific reason of why the connection was closed, is thrown.

Additionally, the connection pool behavior has been improved to ensure
that all connections closed by the server are effectively released from
a pool every time an application tries to acquire a new connection.