WL#14222: rename references of "master" to "trunk"
All links, examples or dummy text related to the main git branch, called
"master" should be renamed to "trunk" instead, since the branch name in
the version control system is also changing.

Nothing else needs to change because there is no other usage of the term
in the source tree.