WL#14257: deprecate EOL Node.js versions
We are officially dropping support for End-of-life Node.js versions. For
now this doesn't mean we are introducing anny breaking change with
regards to JavaScript syntax or core Node.js API usage, but it's
something that might probably happen in the near future.

Starting on 8.0.24, Connector/Node.js will support only Node.js >=12.0.0,
and all the corresponding references to the minimum support version are
being updated.