WL#14406: automate copyright notice header management
Some source files are missing the copyright notice header and others
contain an incorrect header, either because the template has been
updated in the meantime or because the year of either or both the years
of the original and last changes don't match.

This patch introduces tooling that allows to verify issues similar to
these and automatically fix them in the version/source control
lifecycle.