BUG#32136490: fix getDefaultSchema() for undefined schemas
When a default schema is not defined in the connection options, calling
"Session.getDefaultSchema()" returns a valid "Schema" instance in which
we can call any method defined by its API. This is misleading and does
not provide a consistent experience for a JavaScript API.

This patch addresses this issue by ensuring that the
"getDefaultSchema()" method returns "undefined" if the connection
options do not contain a name for the default schema.

Additionally, the associated functional test suite has been revised to
ensure this and all other possible cases are covered.