WL#13733: documentation templating and script
Although documentation in the form of an API reference and extended
tutorials was publicly available, there was no easy way to generate the
respective artifacts, both internally, and for anyone interested in
contributing with improvements and fixes.

Additionally, the process was not flexible enough as to provide
different styles, navigation layouts or even standard placeholders such
as titles and copyright notices, to be replaced by custom values in the
generated output format.

This is being addressed by this patch, which introduces a simple and
portable way to generate the documentation in HTML. The output format
now uses a templating engine that can be easily customizable in the
future.