docs: fix incorrect port in multi-host connection example
One of the examples describing multi-host connection functionality was
misleading because it mentioned an incorrect output, contrary to the
pre-conditions established and described by the example.

This issue was meant to be addressed via a pull request submitted on
GitHub by @asukakenji, but due to some structural changes to the
repository, the pull request was automatically closed before the work
could be merged.

https://github.com/mysql/mysql-connector-nodejs/pull/6

We reached out to the user and suggested submitting a new pull request,
but we didn't hear back.

This commit is meant to express attribution and gratitude. Credit where
credit is due. Thank you @asukakenji.