BUG#32366743: release refused connections back to the pool
If the initial connection to a server is refused, a connection pool
still needs to be usable for a subsequent attempt, given that this might
be a spurious network issue. In that case, when the server eventually
becomes available, an application can acquire a proper connection from
the pool.

This wasn't possible until now, and even if the server became available,
every time an application attempted to acquire a connection from the
pool, it would fail after the number of milliseconds defined by
"queueTimeout" or it would simply block if that value was infinite
(default scenario).

The patch fixes those issues and ensures all refused connections are
available to be re-used by the connection pool in subsequent attempts.