BUG#32188275: add missing schema validation API reference docs
The reference docs for the "Schema" module does not contain important
pointers about schema validation options for both the
"createCollection()" method and the "modifyCollection()" method.

This patch introduces all the important details and fixes the pre-commit
script to ignore common directories used for generated artifacts such as
HTML documentation and code coverage reports.