WL#12430: decouple X DevAPI connections from sessions
The glue code to manage connections to a MySQL server has been since
long tightly-coupled with high-level code used to implement
customer-facing interfaces like the X DevAPI Session module. The fact
that this code does not have a proper place to live also leads, as a
consequence, to a lot of duplication and confusion, mixing of
resposability concerns throughout the codebase.

This makes it extremelly hard to fix issues and introduce enhancements
to all the related high-level aspects such as connection pooling,
multi-host, DNS SRV, autonomous failover and timeout management without
introducing regressions or unintentionally breaking other related
components tied together by the core architecture.

This patch extracts the logic to manage connections to the MySQL server
into a proper "Connection" component and decouples it from the
high-level implementation of the X DevAPI Session blueprint.

As a consequence, other new internal modules have been introduced and
some of the existing components have been refactored, leveraging this
new layout.

The new architecture also leads to additional improvements with regards
to server availability and timeout management alongside overall
connection error handling.

Some changes also involved simplifiying the related setup and glue code
used to create and reproduce the most common connection scenarios and
tests.

Other not-so-related simplifications were introduced as part of the
patch, mainly around removing dead (or not that much useful) code, so
they are not really important to mention.

As part of the changes introduced by the patch, the following bug
reports have also been addressed, with the underlying issues being fixed
and the expected behaviour ensured.

  - BUG#32200234
  - BUG#32205365