tests: fix failing test hooks on Windows
The "afterEach()" test hook for the timeout tests using Unix sockets is
running and failing on Windows. This is only reproducible with the
latest version of the mocha test runner, introduced as part of WL#14258.

The failure itself is related to the fact that the hook tries to close a
server which is not running (since the Unix socket tests are skipped).

In order to prevent this, the hook code has been updated to ensure it
only tries to close the server if it is effectively listening.
Additionally, it ensures to connection can be created while closing the
existing ones.