BUG#32631911: execute metadata callback once per result set
When the "execute()" method supports a push-based cursor in the form of
a callback function used for handling column metadata, such as in the
case of the "Session.sql()" API, that function is currently being called
for each row that is returned as part of the result set.

Such a behavior is pointless, because the column metadata does not
change for a given result set. It can, however, change, when the
statement yields multiple result sets. In any case, processing metadata
when it is not needed only leads to confusion and contributes to
performance degradation.

This patch addresses this situation and introduces changes to ensure the
column metadata callback is called only once for each result set
generated by a statement.