WL#14565: deprecate TLSv1 and TLSv1.1
TLSv1 and TLSv1.1 are now deprecated in the MySQL server. These
deprecated TLS versions are still allowed, and any application that
explicitely wants to use them, will be able to, until the MySQL server
completely removes support for them. In the meantime, those applications
should start migrating to TLSv1.2 or TLSv1.3.

This patch introduces the logic to generate a deprecation warning
message if the negotiated TLS protocol version is lower than TLSv1.2.
This warning message is generated for every standalone connection and
for new and re-created (expired) pool connections. Re-using an idle
connection from the pool will not generate the warning.

The warning message is published in the debug log when an application is
running in debug mode (NODE_DEBUG), otherwise it will be written in the
process stderr stream as 'warning' event, using the same infrastructure
as the one used for deprecation warning messages generated by the core
Node.js engine.

To filter these warning messages in the debug log, the following tags
(among others) can be used:

NODE_DEBUG='connection:tls.version'
NODE_DEBUG='connection:tls.*'
NODE_DEBUG='connection:*'

For more details on how to handle 'warning' events, check the official
Node.js documentation:

https://nodejs.org/docs/v12.0.0/api/process.html#process_event_warning