misc: centralize error and warning infrastructure
Error and warning messages are scattered all over the code base and the
warnings are handled using different mechanisms which leads to some
consistency and maintainability issues.

For instance, in the case of deprecation warnings, there is no
integration with the internal logging framework, and there is no proper
instrumentation in place which is also a cumbersome with regards to
observability.

This patch ensures all errors and warnings are defined using a central
repository and the warnings are properly handled using a single standard
framework, the "process.emitWarning()" API, and are included in the
application log when it runs in debug mode.