BUG#32750927: allow re-use of SQL statement instances
Creating an instance of a SQL statement (containing placeholders) in the
application code, and then trying to execute it multiple times using
different placeholder values results in a "Too many arguments" error
being reported by the server (X Plugin).

This happens because placeholder values are not removed from the list of
arguments passed to the server after each time the statement is
executed, and the new values are appended to that list, leading the
server to reject a list of arguments that ends up containing more values
than the number of placeholders defined in the statement.

Of course the same does not happen when a new statement instance is
created altogether.

  const statement = session.sql('select ?')
  await statement.bind('foo').execute() // works
  await statement.bind('bar').execute() // Error: Too many arguments

Allowing such a behavior also makes the Session.sql() API more
consistent with the CRUD APIs that support server-side prepared
statements.

This patch introduces the changes to allow these SQL statement instances
to be re-used and executed both with the same or with different
placeholder values by freezing each statement after it is executed.