tests: fix JSON debug log content collection issues
The script to collect content from the debug log assumes that each
'data' event generated on the 'stderr' stream corresponds to a single
log entry. That is not always the case and the underlying event emitter
seems to buffer data at times.

Under this assumption, the script tries to create a JSON array
containing JSON objects as items, mapped directly from each entry
extracted from the log. When the 'data' event emits buffered data, there
can be more than one entry available in the resulting content, which
means that bindly converting it to a JSON object will result in a
"SyntaxError" while calling "JSON.parse()".

This patch fixes the script to ensure the content is correctly parsed to
account for scenarios such as the one mentioned above, before feeding
any content to "JSON.parse()" to build the resulting plain JavaScript
object mapping to a single log entry.