tests: less strict expectations involving *INT column length
The "Column.getLength()" API directly returns a value sent by the server
(X Plugin), which is determined by the "max_length" value associated to
a given column data type. In the case, of integer (*INT) and float
types, this value is just advisory. The server is phasing out this kind
of auxiliary display sizes, and that is causing some changes to the
values being originally reported for this type of columns.

This patch updates all the functional tests that contain expectations
about the "Column.getLength()" API in order to make them less strict and
avoid checking for specific values as a mean to avoid issues with these
expected changes in the server.