tests: skip tests failing with deprecated TLS versions
A few functional tests are restricting the list of allowed TLS versions
in order to force the client to negotiate a now deprecated TLS version
(TLSv1 and/or TLSv1.1) with the server.

In theory this is fine and the tests make sense, at least until those
versions stop to be supported by the MySQL server (which is something
that will eventually happen in the future). However, there is a chance
that the underlying platform the application and the Node.js engine are
running on top of, already does not support those TLS versions (for
instance, when using a hardened OpenSSL configuration).

An example where this happens is Ubuntu 20.04.

https://bugs.launchpad.net/ubuntu/+source/gnutls28/+bug/1856428

This means that it does not matter if the MySQL server still supports
the deprecated TLS versions, because it will never be able to negotiate
them on those platforms.

At the moment, we still want to ensure they are being properly treated
as deprecated TLS versions, for the cases where it applies, so we do not
want to remove the tests or somehow change their expectations.

In order to accommodate these needs and prevent the tests from failing
on those specific platforms, we are introducing specific error checks in
the tests to allow them to be "skipped" and effectivelly signal that the
expectations have not been properly verified in those specific cases,
whilst maintaining the same exact set of expectations for the platforms
where they apply.