tests: handle potential warnings when collecting logs
The script used to collect log content for automated functional tests
does not have into account the fact that 'warning' event messages are
also being reported on the stderr stream. This is causing some issues
while trying to parse legitimate log content generated by the tests.

This patch improves the script to ensure the underlying child process
fork is being created using the "--no-warnings" flag, to prevent
'warning' event messages to be reported on stderr and make it easier for
the script to extract the potential JSON content generated by the
package.