BUG#32763298: handle id mismatch on "replaceOne()"
When the replacement document provided as the second argument to the
"Collection.replaceOne()" method contains an "_id" property whose
value does not match the one provided as the first argument, the client
ignores it and replaces any corresponding document in the database or
does nothing if it does not exist.

This is confusing and can ultimately lead to unintended
consequences, such as the case where an application is handling some
kind of document it intends to update making the required changes
inline and somehow providing a wrong value for the "_id" as the first
argument in the method call.

To play it safe, when that happens, and the document effectively
contains a mismatching "_id" property value, the client should report an
error. This patch accordingly implements this change in behavior.