BUG#32820267: replace deprecated Buffer API
Using the "Buffer()" constructor is deprecated in Node.js for some time
because, on older versions, memory allocated for such "Buffer" instances
is not initialized and can contain sensitive data. Although these
instances can be explicitelly initialized using the "fill()" method, the
Node.js core API now includes utility methods to properly create and/or
initialize the underlying allocated memory in a standard fashion. These
are now recommended in favor of the "Buffer()" constructor and using
the later will result in a deprecation warning.

The "Buffer()" constructor is being replaced on a case-by-case scenario
with one of "Buffer.alloc()", "Buffer.allocUnsafe()" and "Buffer.from()"
given the purpose and associated usage conditions. For instance, if the
buffer is supposed to be a view of an existing "ArrayBuffer", we are
using "Buffer.from()", whereas if the buffer is being created from
scratch, we use "Buffer.allocUnsafe()" when the entire buffer is filled
(because it is faster) or "Buffer.alloc()" if not (because it is safer).

A non-comprehensive list of affected components includes:

  - authentication mechanism plugins
  - expression parser partials encoding protobuf types
  - protobug message wrappers encoding and decoding messages
  - additional functional tests handling binary data