BUG#32687374: Upstream floating point number precision
Floating point numbers are being encoded as either X Protocol V_DOUBLE
or V_FLOAT protobuf types based on precision requirements. However, the
X Plugin seems to be losing precision when working with numbers encoded
as V_FLOAT.

Althought it might be better to address this issue at the server side,
there is not any big downside in simply encoding all floating point
numbers in the client-side as V_DOUBLE, which seems to be working
fine.

This patch introduces that workaround and ensures that floating point
numbers represented by a JavaScript number or by literals within an
X DevAPI expression string are always encoded using V_DOUBLE as part of X
Protocol messages sent to the server.