BUG#33175092: fix connection pool parallelism issues
The connection pool implementations has some parallelism issues when it
comes to acquiring connections. When an application creates a pool and
tries to acquire connections in parallel using "Promise.all()", the
internal pool state becomes inconsistent leading to race conditions, and
connections being created outside the scope of the pool.

The issues are particularly troublesome when the pool contains
connections that have been released or have expired.

This patch introduces some changes to the pool implementation and adds
an internal ticketing system to guarantee that requests in parallel are
queued until the internal state is consistent. Once that happens, the
regular rules based on "queueTimeout" continue to apply.