tests: fix default schema in extended functional tests
Due to the changes introduce in the test suite strucutre, some extended
functional tests that relied on the existence of a default schema, were
failing, because the schema was not being created in the same server
instance the test was executing.

This patch introduces some changes to those tests and to the fixture
that allows to create schemas, to ensure they execute in the same host,
thus using the same server instance.