BUG#33175234: error closing standalone connection client
Using a "Client" instance and calling "getSession()" to create a
standalone connection is a bit weird but it is still a valid pattern.
However, in such a scenario, after the connection is created, calling
"Client.close()" yields the following error:

  Error: Cannot close the pool. Maybe it has been destroyed already.

The error message is confusing, since there is no connection pool after
all. Also, there is no valid reason for reporting an error, because, in
the worst-case scenario, nothing should happen, and in the best-case
scenario, the outcome should be functionally equivalent to calling the
"close()" method of the resulting "Session" instance itself.

This patch introduces the changes required to ensure the latter
behavior. So, the following will, from now on, be true:

  client = mysqlx.getClient(/* config */, { pooling: { enabled: false } })
  session = await client.getSession()
  // closing the client instance
  await client.close()
  // is equivalent to closing the session instance
  await session.close()