WL#14810: Remove support for TLSv1 and TLSv1.1
MySQL 8.0.28 does not support neither TLSv1 nor TLSv1.1 anymore. This is
a breaking change for applications that restrict the list of TLS
versions allowed in the client to one or both of these which will, from
now on, not be able to connect to newer MySQL server versions.

These TLS versions have been dreprecated since the 8.0.26 release of
both the official server and clients and are now being effectively
removed on both.

From the client standpoint, this means that starting with the 8.0.28
release, the client will not be able to connect to any MySQL server
instance, if it restricts the list of TLS versions to TLSv1 and/or
TLSv1.1.

To reduce the damage caused by this breaking change, clients will always
attempt to connect to the server and negotiate a TLS version, as long as
a valid one is available in the list (either TLSv1.2 or TLSv1.3), even
if other versions in the list are unknown or unsupported.

This patch introduces the changes required to implement these changes in
behavior.